test: isolate node builtin timer tests
